Remove some code duplication when starting via an Intent
